NFT Market Place added

+++ frontend/src/components/desktop/cards/featuredcoin/FeaturedCoinCards.tsx
@@ -421,7 +421,7 @@ const FeaturedCoinCards = ({ cardData }: any) => {
                   sx={{ alignItems: "center", justifyContent: "center" }}
                   spacing={0}
                 >
-                  {Math.sign(parseInt(cardData?.percent_change_1h)) === -1
+                  {Math.sign(parseFloat(cardData?.percent_change_1h)) === -1
                     ? cardData &&
                       cardData?.percent_change_1h !== null && (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
@@ -435,7 +435,8 @@ const FeaturedCoinCards = ({ cardData }: any) => {
                     variant="caption"
                     sx={{
                       color:
-                        Math.sign(parseInt(cardData?.percent_change_1h)) === -1
+                        Math.sign(parseFloat(cardData?.percent_change_1h)) ===
+                        -1
                           ? "#ff0000"
                           : "#00ff00",
                       fontWeight: 600,
@@ -475,7 +476,7 @@ const FeaturedCoinCards = ({ cardData }: any) => {
                   sx={{ alignItems: "center", justifyContent: "center" }}
                   spacing={0}
                 >
-                  {Math.sign(parseInt(cardData?.percent_change_24h)) === -1
+                  {Math.sign(parseFloat(cardData?.percent_change_24h)) === -1
                     ? cardData &&
                       cardData?.percent_change_24h !== null && (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
@@ -489,7 +490,8 @@ const FeaturedCoinCards = ({ cardData }: any) => {
                     variant="caption"
                     sx={{
                       color:
-                        Math.sign(parseInt(cardData?.percent_change_24h)) === -1
+                        Math.sign(parseFloat(cardData?.percent_change_24h)) ===
+                        -1
                           ? "#ff0000"
                           : "#00ff00",
                       fontWeight: 600,
