+++ frontend/src/components/desktop/htmltable/HtmlTable.tsx
@@ -143,7 +143,7 @@ const HtmlTable = ({ tableData, variant, tableHeader }: any) => {
                     >
                       {data &&
                       data?.pc_24h !== null &&
-                      Math.sign(parseInt(data?.pc_24h)) === -1 ? (
+                      Math.sign(parseFloat(data?.pc_24h)) === -1 ? (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
                       ) : (
                         data?.pc_24h !== null && (
@@ -155,7 +155,7 @@ const HtmlTable = ({ tableData, variant, tableHeader }: any) => {
                         variant="caption"
                         sx={{
                           color:
-                            Math.sign(parseInt(data?.pc_24h)) === -1
+                            Math.sign(parseFloat(data?.pc_24h)) === -1
                               ? "#ff0000"
                               : "#00ff00",
                           fontWeight: 600,
@@ -180,7 +180,7 @@ const HtmlTable = ({ tableData, variant, tableHeader }: any) => {
                     >
                       {data &&
                       data?.pc_7d !== null &&
-                      Math.sign(parseInt(data?.pc_7d)) === -1 ? (
+                      Math.sign(parseFloat(data?.pc_7d)) === -1 ? (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
                       ) : (
                         data?.pc_7d !== null && (
@@ -192,7 +192,7 @@ const HtmlTable = ({ tableData, variant, tableHeader }: any) => {
                         variant="caption"
                         sx={{
                           color:
-                            Math.sign(parseInt(data?.pc_7d)) === -1
+                            Math.sign(parseFloat(data?.pc_7d)) === -1
                               ? "#ff0000"
                               : "#00ff00",
                           fontWeight: 600,
@@ -567,7 +567,7 @@ const HtmlTable = ({ tableData, variant, tableHeader }: any) => {
                     >
                       {data &&
                       data?.percent_change_1h !== null &&
-                      Math.sign(parseInt(data?.percent_change_1h)) === -1 ? (
+                      Math.sign(parseFloat(data?.percent_change_1h)) === -1 ? (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
                       ) : (
                         data?.percent_change_1h !== null && (
@@ -579,7 +579,8 @@ const HtmlTable = ({ tableData, variant, tableHeader }: any) => {
                         variant="caption"
                         sx={{
                           color:
-                            Math.sign(parseInt(data?.percent_change_1h)) === -1
+                            Math.sign(parseFloat(data?.percent_change_1h)) ===
+                            -1
                               ? "#ff0000"
                               : "#00ff00",
                           fontWeight: 600,
@@ -610,7 +611,7 @@ const HtmlTable = ({ tableData, variant, tableHeader }: any) => {
                     >
                       {data &&
                       data?.percent_change_24h !== null &&
-                      Math.sign(parseInt(data?.percent_change_24h)) === -1 ? (
+                      Math.sign(parseFloat(data?.percent_change_24h)) === -1 ? (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
                       ) : (
                         data?.percent_change_24h !== null && (
@@ -622,7 +623,8 @@ const HtmlTable = ({ tableData, variant, tableHeader }: any) => {
                         variant="caption"
                         sx={{
                           color:
-                            Math.sign(parseInt(data?.percent_change_24h)) === -1
+                            Math.sign(parseFloat(data?.percent_change_24h)) ===
+                            -1
                               ? "#ff0000"
                               : "#00ff00",
                           fontWeight: 600,
