+++ frontend/src/components/mobile/htmltable/MobileHtmlTable.tsx
@@ -140,7 +140,7 @@ const MobileHtmlTable = ({ tableData, variant, tableHeader }: any) => {
                     >
                       {data &&
                       data?.pc_24h !== null &&
-                      Math.sign(parseInt(data?.pc_24h)) === -1 ? (
+                      Math.sign(parseFloat(data?.pc_24h)) === -1 ? (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
                       ) : (
                         data?.pc_24h !== null && (
@@ -152,7 +152,7 @@ const MobileHtmlTable = ({ tableData, variant, tableHeader }: any) => {
                         variant="caption"
                         sx={{
                           color:
-                            Math.sign(parseInt(data?.pc_24h)) === -1
+                            Math.sign(parseFloat(data?.pc_24h)) === -1
                               ? "#ff0000"
                               : "#00ff00",
                           fontWeight: 600,
@@ -177,7 +177,7 @@ const MobileHtmlTable = ({ tableData, variant, tableHeader }: any) => {
                     >
                       {data &&
                       data?.pc_7d !== null &&
-                      Math.sign(parseInt(data?.pc_7d)) === -1 ? (
+                      Math.sign(parseFloat(data?.pc_7d)) === -1 ? (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
                       ) : (
                         data?.pc_7d !== null && (
@@ -189,7 +189,7 @@ const MobileHtmlTable = ({ tableData, variant, tableHeader }: any) => {
                         variant="caption"
                         sx={{
                           color:
-                            Math.sign(parseInt(data?.pc_7d)) === -1
+                            Math.sign(parseFloat(data?.pc_7d)) === -1
                               ? "#ff0000"
                               : "#00ff00",
                           fontWeight: 600,
@@ -564,7 +564,7 @@ const MobileHtmlTable = ({ tableData, variant, tableHeader }: any) => {
                     >
                       {data &&
                       data?.percent_change_1h !== null &&
-                      Math.sign(parseInt(data?.percent_change_1h)) === -1 ? (
+                      Math.sign(parseFloat(data?.percent_change_1h)) === -1 ? (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
                       ) : (
                         data?.percent_change_1h !== null && (
@@ -576,7 +576,8 @@ const MobileHtmlTable = ({ tableData, variant, tableHeader }: any) => {
                         variant="caption"
                         sx={{
                           color:
-                            Math.sign(parseInt(data?.percent_change_1h)) === -1
+                            Math.sign(parseFloat(data?.percent_change_1h)) ===
+                            -1
                               ? "#ff0000"
                               : "#00ff00",
                           fontWeight: 600,
@@ -607,7 +608,7 @@ const MobileHtmlTable = ({ tableData, variant, tableHeader }: any) => {
                     >
                       {data &&
                       data?.percent_change_24h !== null &&
-                      Math.sign(parseInt(data?.percent_change_24h)) === -1 ? (
+                      Math.sign(parseFloat(data?.percent_change_24h)) === -1 ? (
                         <ArrowDropDownIcon sx={{ color: "#ff0000" }} />
                       ) : (
                         data?.percent_change_24h !== null && (
@@ -619,7 +620,8 @@ const MobileHtmlTable = ({ tableData, variant, tableHeader }: any) => {
                         variant="caption"
                         sx={{
                           color:
-                            Math.sign(parseInt(data?.percent_change_24h)) === -1
+                            Math.sign(parseFloat(data?.percent_change_24h)) ===
+                            -1
                               ? "#ff0000"
                               : "#00ff00",
                           fontWeight: 600,
